refactor(core): split core_api end
