+++ core/src/api/tokenize.rs
@@ -1,0 +1,112 @@
+use crate::providers::provider::{provider, ProviderID};
+use crate::run;
+use crate::utils::{error_response, APIResponse};
+use axum::Json;
+use http::StatusCode;
+use serde_json::json;
+
+#[derive(serde::Deserialize)]
+pub struct TokenizePayload {
+    text: String,
+    provider_id: ProviderID,
+    model_id: String,
+    credentials: Option<run::Credentials>,
+}
+
+pub async fn tokenize(Json(payload): Json<TokenizePayload>) -> (StatusCode, Json<APIResponse>) {
+    let mut llm = provider(payload.provider_id).llm(payload.model_id);
+
+    // If we received credentials we initialize the llm with them.
+    match payload.credentials {
+        Some(c) => {
+            match llm.initialize(c.clone()).await {
+                Err(e) => {
+                    return error_response(
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        "internal_server_error",
+                        "Failed to initialize LLM",
+                        Some(e),
+                    );
+                }
+                Ok(()) => (),
+            };
+        }
+        None => (),
+    }
+
+    match llm.tokenize(vec![payload.text]).await {
+        Err(e) => error_response(
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "internal_server_error",
+            "Failed to tokenize text",
+            Some(e),
+        ),
+        Ok(mut res) => match res.pop() {
+            None => error_response(
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "internal_server_error",
+                "Failed to tokenize text",
+                None,
+            ),
+            Some(tokens) => (
+                StatusCode::OK,
+                Json(APIResponse {
+                    error: None,
+                    response: Some(json!({
+                        "tokens": tokens,
+                    })),
+                }),
+            ),
+        },
+    }
+}
+
+#[derive(serde::Deserialize)]
+pub struct TokenizeBatchPayload {
+    texts: Vec<String>,
+    provider_id: ProviderID,
+    model_id: String,
+    credentials: Option<run::Credentials>,
+}
+
+pub async fn tokenize_batch(
+    Json(payload): Json<TokenizeBatchPayload>,
+) -> (StatusCode, Json<APIResponse>) {
+    let mut llm = provider(payload.provider_id).llm(payload.model_id);
+
+    // If we received credentials we initialize the llm with them.
+    match payload.credentials {
+        Some(c) => {
+            match llm.initialize(c.clone()).await {
+                Err(e) => {
+                    return error_response(
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        "internal_server_error",
+                        "Failed to initialize LLM",
+                        Some(e),
+                    );
+                }
+                Ok(()) => (),
+            };
+        }
+        None => (),
+    }
+
+    match llm.tokenize(payload.texts).await {
+        Err(e) => error_response(
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "internal_server_error",
+            "Failed to tokenize text",
+            Some(e),
+        ),
+        Ok(res) => (
+            StatusCode::OK,
+            Json(APIResponse {
+                error: None,
+                response: Some(json!({
+                    "tokens": res,
+                })),
+            }),
+        ),
+    }
+}
